برداشتن فیلتر پرسنل به خاطر انتخاب بیمار در ارجاع تکمیل ارجاع

+++ src/main/java/com/bzamani/framework/service/impl/core/personel/PersonelService.java
@@ -1,9 +1,12 @@
 package com.bzamani.framework.service.impl.core.personel;
 
+import com.bzamani.framework.common.utility.SecurityUtility;
 import com.bzamani.framework.model.core.personel.Personel;
 import com.bzamani.framework.repository.core.personel.IPersonelRepository;
 import com.bzamani.framework.service.core.file.IFileAttachmentService;
+import com.bzamani.framework.service.core.organization.IOrganizationService;
 import com.bzamani.framework.service.core.personel.IPersonelService;
+import com.bzamani.framework.service.core.user.IUserService;
 import com.bzamani.framework.service.impl.core.GenericService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
@@ -27,6 +30,12 @@ public class PersonelService extends GenericService<Personel, Long> implements I
     @Autowired
     private IFileAttachmentService iFileAttachmentService;
 
+    @Autowired
+    private IUserService iUserService;
+
+    @Autowired
+    private IOrganizationService iOrganizationService;
+
     @Override
     protected JpaRepository<Personel, Long> getGenericRepo() {
         return iPersonelRepository;
@@ -34,7 +43,11 @@ public class PersonelService extends GenericService<Personel, Long> implements I
 
     @Override
     @Transactional
-    public Personel save(Personel personel) {
+    public Personel checkAndSave(Personel personel) throws Exception {
+        if (!iOrganizationService.userHaveAccessToOrganization(
+                iUserService.findUserByUsernameEquals(SecurityUtility.getAuthenticatedUser().getUsername()).getId(), personel.getOrganization().getId()))
+            throw new Exception("امکان ویرایش اطلاعات این فرد برای شما که به واحد سازمانی آن دسترسی ندارید وجود ندارد.");
+
         personel.setFileCode(personel.getFileCode() == null || personel.getFileCode().length() == 0 ? null : personel.getFileCode());
         String oldFileCode = null;
         String newFileCode = personel.getFileCode();
@@ -46,9 +59,13 @@ public class PersonelService extends GenericService<Personel, Long> implements I
         return super.save(personel);
     }
 
-    @Override
     @Transactional
-    public boolean deleteByEntityId(Long id) {
+    @Override
+    public boolean checkAndDelete(Long id) throws Exception {
+        Personel personel = loadByEntityId(id);
+        if (!iOrganizationService.userHaveAccessToOrganization(
+                iUserService.findUserByUsernameEquals(SecurityUtility.getAuthenticatedUser().getUsername()).getId(), personel.getOrganization().getId()))
+            throw new Exception("امکان حذف فرد برای شما که به واحد سازمانی این فرد دسترسی ندارید وجود ندارد.");
         if (loadByEntityId(id).getFileCode() != null)
             iFileAttachmentService.finalizeNewAndDeleteOldAttachment(null, loadByEntityId(id).getFileCode());
         return super.deleteByEntityId(id);
@@ -74,6 +91,35 @@ public class PersonelService extends GenericService<Personel, Long> implements I
                 lastname,
                 mobile,
                 organizationId, pagingSort);
+        personels = pageTuts.getContent();
+        Map<String, Object> response = new HashMap<>();
+        response.put("entityList", personels);
+        response.put("currentPage", pageTuts.getNumber());
+        response.put("totalRecords", pageTuts.getTotalElements());
+        response.put("totalPages", pageTuts.getTotalPages());
+        return response;
+    }
+
+    @Override
+    public Map<String, Object> searchPersonelAuthorize(String firstname,
+                                                       String lastname,
+                                                       String mobile,
+                                                       Long organizationId, int page, int size, String[] sort) {
+        List<Sort.Order> orders = new ArrayList<Sort.Order>();
+        if (sort[0].contains(",")) {
+            for (String sortOrder : sort) {
+                String[] _sort = sortOrder.split(",");
+                orders.add(new Sort.Order(getSortDirection(_sort[1]), _sort[0]));
+            }
+        } else {
+            orders.add(new Sort.Order(getSortDirection(sort[1]), sort[0]));
+        }
+        List<Personel> personels = new ArrayList<Personel>();
+        Pageable pagingSort = PageRequest.of(page, size, Sort.by(orders));
+        Page<Personel> pageTuts = iPersonelRepository.searchPersonelAuthorize(firstname,
+                lastname,
+                mobile,
+                organizationId, iUserService.findUserByUsernameEquals(SecurityUtility.getAuthenticatedUser().getUsername()).getId(), pagingSort);
         personels = pageTuts.getContent();
         Map<String, Object> response = new HashMap<>();
         response.put("entityList", personels);
